+++ modules/traefik/internal.tf
@@ -124,9 +124,15 @@ resource "aws_lb_listener" "internal_http" {
   port              = "80"
   protocol          = "HTTP"
 
+  # Redirect to HTTPS
   default_action {
-    target_group_arn = "${aws_lb_target_group.internal.arn}"
-    type             = "forward"
+    type = "redirect"
+
+    redirect {
+      protocol    = "HTTPS"
+      port        = 443
+      status_code = "HTTP_301"
+    }
   }
 }